cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imdb"
version = "0.1.0"
description = "CLI para consultar datos de IMDb"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
]

[project.scripts]
imdb = "imdb:cli"
